refactor: extract trait-based backend module with local implementation (#6)
Refactor the monolithic broker.rs and client.rs into a trait-based
backend module. The Backend trait defines serve() and send_request()
methods, with LocalBackend as the concrete implementation using Unix
domain sockets.

- Add src/backend/mod.rs with Backend trait and create_backend() factory
- Add src/backend/local.rs with all broker state, server, and client logic
- Add UnknownBackend error variant to DispatchError
- Wire config.backend field to create_backend() in main.rs
- Remove src/broker.rs and src/client.rs (logic moved, not changed)
- All 56 tests pass, no behavior changes
