+++ Cargo.toml
@@ -18,6 +18,7 @@ semver = "1"
 uuid = { version = "1", features = ["v4"] }
 serde_json = "1"
 dirs = "6"
+async-trait = "0.1.89"
 
 [profile.release]
 strip = true
